refactor: move AboutIntoCard into ArchivedComponents
