+++ src/pages/about-us/partners.tsx
@@ -4,7 +4,7 @@ import { FC } from "react";
 import CMSClient from "../../node-lib/cms";
 import { AboutPartnersPage } from "../../common-lib/cms-types";
 import AboutContactCard from "../../components/GenericPagesComponents/AboutContactCard";
-import AboutIntroCard from "../../components/AboutIntoCard/AboutIntroCard";
+import AboutIntroCard from "../../components/ArchivedComponents/AboutIntoCard/AboutIntroCard";
 import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
 import AboutUsSummaryCard from "../../components/pages/AboutUs/AboutUsSummaryCard";
 import getPageProps from "../../node-lib/getPageProps";
